Separated component to different file
